cv/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17
google-generativeai==0.8.3
pillow==11.0.0
pydantic==2.10.0
asyncio==3.4.3
tenacity==8.2.2